requirements.txt
DateTime==5.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
